+++ src/utility/cardCovers.js
@@ -1,13 +1,13 @@
-import cover1 from '../assets/celtic-gem.png'
-import cover2 from '../assets/diamond-grasp.png'
-import cover3 from '../assets/emerald-grasp.png'
-import cover4 from '../assets/eye-clasp.png'
-import cover5 from '../assets/eye-flower.png'
-import cover6 from '../assets/hand-gem-pandelum.png'
-import cover7 from '../assets/lunar-flower.png'
-import cover8 from '../assets/moon-pandelum.png'
-import cover9 from '../assets/saturn-pandelum.png'
-import cover10 from '../assets/shiny-gem.png'
-import cover11 from '../assets/sun-lunar-flower.png'
+import cover1 from '../assets/images/celtic-gem.png'
+import cover2 from '../assets/images/diamond-grasp.png'
+import cover3 from '../assets/images/emerald-grasp.png'
+import cover4 from '../assets/images/eye-clasp.png'
+import cover5 from '../assets/images/eye-flower.png'
+import cover6 from '../assets/images/hand-gem-pandelum.png'
+import cover7 from '../assets/images/lunar-flower.png'
+import cover8 from '../assets/images/moon-pandelum.png'
+import cover9 from '../assets/images/saturn-pandelum.png'
+import cover10 from '../assets/images/shiny-gem.png'
+import cover11 from '../assets/images/sun-lunar-flower.png'
 
 export const covers = [ cover1, cover2, cover3, cover4, cover5, cover6, cover7, cover8, cover9, cover10, cover11 ]
